Minor fixes after merging branches

+++ Maximum flow in the graph/Graph.h
@@ -13,7 +13,7 @@ struct Graph
 		}
 	}
 
-	const int V;
+	const size_t V;
 	std::vector<std::vector<int>> adj;
 };
